refactor: update judge

+++ services/judge/src/main/java/cloud/oj/judge/entity/Problem.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Problem {
-    private long timeout;
+    private int timeout;
     private int memoryLimit;
     private int outputLimit;
     private int score;
